# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "grtblog 后端接口",
        "title": "grtblog API v2",
        "contact": {},
        "version": "2.0.0"
    },
    "basePath": "/api/v2",
    "paths": {
...
        "/articles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "tagId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ArticleListResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "description": "创建文章参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CreateArticleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ArticleResp"
                        }
                    }
                }
            }
        },
        "/articles/short/{shortUrl}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "根据短链接获取文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接",
                        "name": "shortUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ArticleResp"
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ArticleResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新文章参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateArticleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ArticleResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/articles/{id}/latest": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "校验文章是否最新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文章版本校验参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CheckArticleLatestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.CheckArticleLatestResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contract.ArticleListItemResp": {
            "type": "object",
            "properties": {
                "authorName": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryShortUrl": {
                    "type": "string"
                },
                "comments": {
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHot": {
                    "type": "boolean"
                },
                "isOriginal": {
                    "type": "boolean"
                },
                "isTop": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "shortUrl": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "contract.ArticleListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.ArticleListItemResp"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "contract.ArticleResp": {
            "type": "object",
            "properties": {
                "aiSummary": {
                    "type": "string"
                },
                "authorId": {
                    "type": "integer"
                },
                "categoryId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "contentHash": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHot": {
                    "type": "boolean"
                },
                "isOriginal": {
                    "type": "boolean"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "isTop": {
                    "type": "boolean"
                },
                "leadIn": {
                    "type": "string"
                },
                "metrics": {
                    "$ref": "#/definitions/contract.MetricsResp"
                },
                "shortUrl": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TagResp"
                    }
                },
                "title": {
                    "type": "string"
                },
                "toc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "contract.CheckArticleLatestReq": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                }
            }
        },
        "contract.CheckArticleLatestResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "contentHash": {
                    "type": "string"
                },
                "latest": {
                    "type": "boolean"
                },
                "leadIn": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                }
            }
        },
...
        "contract.CreateArticleReq": {
            "type": "object",
            "required": [
                "content",
                "isPublished",
                "title"
            ],
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "可以自定义发布时间",
                    "type": "string"
                },
                "isHot": {
                    "type": "boolean"
                },
                "isOriginal": {
                    "type": "boolean"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "isTop": {
                    "type": "boolean"
                },
                "leadIn": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tagIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "contract.MetricsResp": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
        "contract.TOCNode": {
            "type": "object",
            "properties": {
                "anchor": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "contract.TagResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "contract.UpdateArticleReq": {
            "type": "object",
            "required": [
                "content",
                "shortUrl",
                "title"
            ],
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "isHot": {
                    "type": "boolean"
                },
                "isOriginal": {
                    "type": "boolean"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "isTop": {
                    "type": "boolean"
                },
                "leadIn": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tagIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}